requirements.txt
pybluez
unidecode==1.0.23
requests==2.21.0
geocoder==1.38.1
pillow==5.4.1